@@ -10,7 +10,7 @@
         }
     ],
     "require": {
-        "php": ">=5.4.0",
+        "php": ">=7.2.0",
         "ext-curl": "*",
         "ext-dom": "*",
         "ext-hash": "*",
@@ -18,7 +18,7 @@
         "ext-xmlwriter": "*"
     },
     "require-dev": {
-        "phpunit/phpunit": "3.7.*"
+        "phpunit/phpunit": "^7.5"
     },
     "autoload": {
         "psr-0": {
