@@ -10,7 +10,7 @@ class CustomerTest extends Setup
 {
     public function testGet_givesErrorIfInvalidProperty()
     {
-        $this->setExpectedException('PHPUnit_Framework_Error', 'Undefined property on Braintree\Customer: foo');
+        $this->expectException('PHPUnit\Framework\Error\Error', 'Undefined property on Braintree\Customer: foo');
         $c = Braintree\Customer::factory([]);
         $c->foo;
     }
@@ -49,13 +49,13 @@ class CustomerTest extends Setup
 
     public function testFindErrorsOnBlankId()
     {
-        $this->setExpectedException('InvalidArgumentException');
+        $this->expectException('InvalidArgumentException');
         Braintree\Customer::find('');
     }
 
     public function testFindErrorsOnWhitespaceId()
     {
-        $this->setExpectedException('InvalidArgumentException');
+        $this->expectException('InvalidArgumentException');
         Braintree\Customer::find('\t');
     }
 }
