@@ -16,7 +16,6 @@ namespace Braintree;
  * @property-read \Braintree\AndroidPayCard[] $androidPayCards
  * @property-read \Braintree\AmexExpressCheckoutCard[] $amexExpressCheckoutCards
  * @property-read \Braintree\ApplePayCard[] $applePayCards
- * @property-read \Braintree\CoinbaseAccount[] $coinbaseAccounts
  * @property-read string $company
  * @property-read \DateTime $createdAt
  * @property-read \Braintree\CreditCard[] $creditCards
@@ -77,25 +76,6 @@ class Customer extends Base
     public static function createNoValidate($attribs = [])
     {
         return Configuration::gateway()->customer()->createNoValidate($attribs);
-    }
-
-    /**
-     * @deprecated since version 2.3.0
-     * @param string $queryString
-     * @return Result\Successful
-     */
-    public static function createFromTransparentRedirect($queryString)
-    {
-        return Configuration::gateway()->customer()->createFromTransparentRedirect($queryString);
-    }
-
-    /**
-     * @deprecated since version 2.3.0
-     * @return string
-     */
-    public static function createCustomerUrl()
-    {
-        return Configuration::gateway()->customer()->createCustomerUrl();
     }
 
     /**
@@ -200,27 +180,6 @@ class Customer extends Base
         return Configuration::gateway()->customer()->updateNoValidate($customerId, $attributes);
     }
 
-    /**
-     *
-     * @deprecated since version 2.3.0
-     * @return string
-     */
-    public static function updateCustomerUrl()
-    {
-        return Configuration::gateway()->customer()->updateCustomerUrl();
-    }
-
-    /**
-     *
-     * @deprecated since version 2.3.0
-     * @param string $queryString
-     * @return Result\Successful|Result\Error
-     */
-    public static function updateFromTransparentRedirect($queryString)
-    {
-        return Configuration::gateway()->customer()->updateFromTransparentRedirect($queryString);
-    }
-
     /* instance methods */
 
     /**
@@ -250,14 +209,6 @@ class Customer extends Base
             }
         }
         $this->_set('creditCards', $creditCardArray);
-
-        $coinbaseAccountArray = [];
-        if (isset($customerAttribs['coinbaseAccounts'])) {
-            foreach ($customerAttribs['coinbaseAccounts'] AS $coinbaseAccount) {
-                $coinbaseAccountArray[] = CoinbaseAccount::factory($coinbaseAccount);
-            }
-        }
-        $this->_set('coinbaseAccounts', $coinbaseAccountArray);
 
         $paypalAccountArray = [];
         if (isset($customerAttribs['paypalAccounts'])) {
@@ -335,7 +286,6 @@ class Customer extends Base
             $this->creditCards,
             $this->paypalAccounts,
             $this->applePayCards,
-            $this->coinbaseAccounts,
             $this->androidPayCards,
             $this->amexExpressCheckoutCards,
             $this->venmoAccounts,
@@ -344,6 +294,12 @@ class Customer extends Base
             $this->samsungPayCards,
             $this->usBankAccounts
         ));
+
+        $customFields = [];
+        if (isset($customerAttribs['customFields'])) {
+            $customFields = $customerAttribs['customFields'];
+        }
+        $this->_set('customFields', $customFields);
     }
 
     /**
@@ -366,18 +322,6 @@ class Customer extends Base
     public function isEqual($otherCust)
     {
         return !($otherCust instanceof Customer) ? false : $this->id === $otherCust->id;
-    }
-
-    /**
-     * returns an array containt all of the customer's payment methods
-     *
-     * @deprecated since version 3.1.0 - use the paymentMethods property directly
-     *
-     * @return array
-     */
-    public function paymentMethods()
-    {
-        return $this->paymentMethods;
     }
 
     /**
