@@ -191,11 +191,11 @@ class Http
         }
 
         curl_close($curl);
-        if ($this->_config->sslOn()) {
-            if ($httpStatus == 0) {
-                throw new Exception\SSLCertificate($error, $error_code);
-            }
-        } else if ($error_code) {
+        if ($this->_config->sslOn() && $error_code == 35) {
+            throw new Exception\SSLCertificate($error, $error_code);
+        }
+
+        if ($error_code) {
             throw new Exception\Connection($error, $error_code);
         }
 
