@@ -17,7 +17,7 @@ class MerchantAccountGateway
 
     public function create($attribs)
     {
-        Util::verifyKeys(self::detectSignature($attribs), $attribs);
+        Util::verifyKeys(self::createSignature(), $attribs);
         return $this->_doCreate('/merchant_accounts/create_via_api', ['merchant_account' => $attribs]);
     }
 
@@ -36,16 +36,6 @@ class MerchantAccountGateway
     {
         Util::verifyKeys(self::updateSignature(), $attributes);
         return $this->_doUpdate('/merchant_accounts/' . $merchant_account_id . '/update_via_api', ['merchant_account' => $attributes]);
-    }
-
-    public static function detectSignature($attribs)
-    {
-        if (isset($attribs['applicantDetails'])) {
-            trigger_error("DEPRECATED: Passing applicantDetails to create is deprecated. Please use individual, business, and funding", E_USER_NOTICE);
-            return self::createDeprecatedSignature();
-        } else {
-            return self::createSignature();
-        }
     }
 
     public static function updateSignature()
@@ -116,31 +106,6 @@ class MerchantAccountGateway
             ['individual' => $individualSignature],
             ['funding' => $fundingSignature],
             ['business' => $businessSignature]
-        ];
-    }
-
-    public static function createDeprecatedSignature()
-    {
-        $applicantDetailsAddressSignature = ['streetAddress', 'postalCode', 'locality', 'region'];
-        $applicantDetailsSignature = [
-            'companyName',
-            'firstName',
-            'lastName',
-            'email',
-            'phone',
-            'dateOfBirth',
-            'ssn',
-            'taxId',
-            'routingNumber',
-            'accountNumber',
-            ['address' => $applicantDetailsAddressSignature]
-        ];
-
-        return [
-            ['applicantDetails' =>  $applicantDetailsSignature],
-            'id',
-            'tosAccepted',
-            'masterMerchantAccountId'
         ];
     }
 
