@@ -60,37 +60,6 @@ class TransactionGateway
         $result = $this->create($attribs);
         return Util::returnObjectOrThrowException(__CLASS__, $result);
     }
-    /**
-     *
-     * @deprecated since version 2.3.0
-     * @access public
-     * @param array $attribs
-     * @return object
-     */
-    public function createFromTransparentRedirect($queryString)
-    {
-        trigger_error("DEPRECATED: Please use TransparentRedirectRequest::confirm", E_USER_NOTICE);
-        $params = TransparentRedirect::parseAndValidateQueryString(
-                $queryString
-        );
-        return $this->_doCreate(
-                '/transactions/all/confirm_transparent_redirect_request',
-                ['id' => $params['id']]
-        );
-    }
-    /**
-     *
-     * @deprecated since version 2.3.0
-     * @access public
-     * @param none
-     * @return string
-     */
-    public function createTransactionUrl()
-    {
-        trigger_error("DEPRECATED: Please use TransparentRedirectRequest::url", E_USER_NOTICE);
-        return $this->_config->baseUrl() . $this->_config->merchantPath() .
-                '/transactions/all/create_via_transparent_redirect_request';
-    }
 
     public static function cloneSignature()
     {
@@ -184,10 +153,6 @@ class TransactionGateway
                     ['threeDSecure' =>
                         ['required']
                     ],
-                    # TODO: Snake case version included for backwards compatiblity. Remove in the next major version
-                    ['three_d_secure' =>
-                        ['required']
-                    ],
                     ['paypal' =>
                         [
                             'payeeId',
@@ -207,8 +172,6 @@ class TransactionGateway
                     ],
                     ['venmo' =>
                         [
-                            # TODO: Snake case version included for backwards compatiblity. Remove in the next major version
-                            'profile_id',
                             'profileId'
                         ]
                     ]
@@ -217,9 +180,6 @@ class TransactionGateway
             ['customFields' => ['_anyKey_']],
             ['descriptor' => ['name', 'phone', 'url']],
             ['paypalAccount' => ['payeeId', 'payeeEmail', 'payerId', 'paymentId']],
-            # TODO: Snake case version included for backwards compatiblity. Remove in the next major version
-            ['apple_pay_card' => ['number', 'cardholder_name', 'cryptogram', 'expiration_month', 'expiration_year', 'eci_indicator']], 
-
             ['applePayCard' => ['number', 'cardholderName', 'cryptogram', 'expirationMonth', 'expirationYear', 'eciIndicator']],
             ['industry' =>
                 ['industryType',
@@ -390,7 +350,7 @@ class TransactionGateway
 
             return new ResourceCollection($response, $pager);
         } else {
-            throw new Exception\DownForMaintenance();
+            throw new Exception\Timeout();
         }
     }
 
@@ -410,7 +370,7 @@ class TransactionGateway
                 'transaction'
             );
         } else {
-            throw new Exception\DownForMaintenance();
+            throw new Exception\Timeout();
         }
     }
 
