@@ -31,7 +31,7 @@ use Braintree\Instance;
  */
 class TransactionLineItem extends Instance
 {
-    // LineItem Kinds
+    // TransactionLineItem Kinds
     const CREDIT = 'credit';
     const DEBIT = 'debit';
 
@@ -51,5 +51,3 @@ class TransactionLineItem extends Instance
     }
 }
 class_alias('Braintree\TransactionLineItem', 'Braintree_TransactionLineItem');
-class_alias('Braintree\TransactionLineItem', 'Braintree\Transaction\LineItem');
-class_alias('Braintree\TransactionLineItem', 'Braintree_Transaction_LineItem');
