@@ -61,6 +61,9 @@ class Util
         case 404:
             throw new Exception\NotFound();
             break;
+        case 408;
+            throw new Exception\RequestTimeout();
+            break;
         case 426:
             throw new Exception\UpgradeRequired();
             break;
@@ -70,8 +73,8 @@ class Util
         case 500:
             throw new Exception\ServerError();
             break;
-        case 503:
-            throw new Exception\DownForMaintenance();
+        case 504;
+            throw new Exception\GatewayTimeout();
             break;
         default:
             throw new Exception\Unexpected('Unexpected HTTP_RESPONSE #' . $statusCode);
@@ -114,7 +117,7 @@ class Util
             } else if ($errorClass == "INTERNAL") {
                 throw new Exception\ServerError();
             } else if ($errorClass == "SERVICE_AVAILABILITY") {
-                throw new Exception\DownForMaintenance();
+                throw new Exception\ServiceUnavailable();
             } else {
                 throw new Exception\Unexpected('Unexpected exception ' . $message);
             }
