@@ -488,7 +488,7 @@ class CreditCardTest extends Setup
     public function testCreateNoValidate_throwsIfValidationsFail()
     {
 
-        $this->setExpectedException('Braintree\Exception\ValidationsFailed');
+        $this->expectException('Braintree\Exception\ValidationsFailed');
         $customer = Braintree\Customer::createNoValidate();
         Braintree\CreditCard::createNoValidate([
             'expirationDate' => 'invalid',
@@ -509,153 +509,6 @@ class CreditCardTest extends Setup
         $this->assertEquals('5100', $creditCard->last4);
         $this->assertEquals('Cardholder', $creditCard->cardholderName);
         $this->assertEquals('05/2012', $creditCard->expirationDate);
-    }
-
-    public function testCreateFromTransparentRedirect()
-    {
-        Test\Helper::suppressDeprecationWarnings();
-        $customer = Braintree\Customer::createNoValidate();
-        $queryString = $this->createCreditCardViaTr(
-            [
-                'credit_card' => [
-                    'number' => '5105105105105100',
-                    'expiration_date' => '05/12'
-                ]
-            ],
-            [
-                'creditCard' => [
-                    'customerId' => $customer->id
-                ]
-            ]
-        );
-        $result = Braintree\CreditCard::createFromTransparentRedirect($queryString);
-        $this->assertTrue($result->success);
-        $this->assertEquals('510510', $result->creditCard->bin);
-        $this->assertEquals('5100', $result->creditCard->last4);
-        $this->assertEquals('05/2012', $result->creditCard->expirationDate);
-    }
-
-    public function testCreateFromTransparentRedirect_withDefault()
-    {
-        Test\Helper::suppressDeprecationWarnings();
-        $customer = Braintree\Customer::createNoValidate();
-        $queryString = $this->createCreditCardViaTr(
-            [
-                'credit_card' => [
-                    'number' => '5105105105105100',
-                    'expiration_date' => '05/12',
-                    'options' => ['make_default' => true]
-                ]
-            ],
-            [
-                'creditCard' => [
-                    'customerId' => $customer->id
-                ]
-            ]
-        );
-        $result = Braintree\CreditCard::createFromTransparentRedirect($queryString);
-        $this->assertTrue($result->creditCard->isDefault());
-    }
-
-    public function testUpdateFromTransparentRedirect()
-    {
-        $customer = Braintree\Customer::createNoValidate();
-        $creditCard = Braintree\CreditCard::createNoValidate([
-            'customerId' => $customer->id,
-            'number' => '5105105105105100',
-            'expirationDate' => '05/12'
-        ]);
-        $queryString = $this->updateCreditCardViaTr(
-            [
-                'credit_card' => [
-                    'number' => '4111111111111111',
-                    'expiration_date' => '01/11'
-                ]
-            ],
-            ['paymentMethodToken' => $creditCard->token]
-        );
-        $result = Braintree\CreditCard::updateFromTransparentRedirect($queryString);
-        $this->assertTrue($result->success);
-        $this->assertEquals('411111', $result->creditCard->bin);
-        $this->assertEquals('1111', $result->creditCard->last4);
-        $this->assertEquals('01/2011', $result->creditCard->expirationDate);
-    }
-
-    public function testUpdateFromTransparentRedirect_withDefault()
-    {
-        $customer = Braintree\Customer::createNoValidate();
-        $card1 = Braintree\CreditCard::createNoValidate([
-            'customerId' => $customer->id,
-            'number' => '5105105105105100',
-            'expirationDate' => '05/12'
-        ]);
-        $card2 = Braintree\CreditCard::createNoValidate([
-            'customerId' => $customer->id,
-            'number' => '5105105105105100',
-            'expirationDate' => '05/12'
-        ]);
-        $this->assertFalse($card2->isDefault());
-
-        $queryString = $this->updateCreditCardViaTr(
-            [
-                'credit_card' => [
-                    'options' => [
-                        'make_default' => true
-                    ]
-                ]
-            ],
-            ['paymentMethodToken' => $card2->token]
-        );
-        $result = Braintree\CreditCard::updateFromTransparentRedirect($queryString);
-        $this->assertFalse(Braintree\CreditCard::find($card1->token)->isDefault());
-        $this->assertTrue(Braintree\CreditCard::find($card2->token)->isDefault());
-    }
-
-    public function testUpdateFromTransparentRedirect_andUpdateExistingBillingAddress()
-    {
-        $customer = Braintree\Customer::createNoValidate();
-        $card = Braintree\CreditCard::createNoValidate([
-            'customerId' => $customer->id,
-            'number' => '5105105105105100',
-            'expirationDate' => '05/12',
-            'billingAddress' => [
-                'firstName' => 'Drew',
-                'lastName' => 'Smith',
-                'company' => 'Smith Co.',
-                'streetAddress' => '123 Old St',
-                'extendedAddress' => 'Suite 101',
-                'locality' => 'Chicago',
-                'region' => 'IL',
-                'postalCode' => '60622',
-                'countryName' => 'United States of America'
-            ]
-        ]);
-
-        $queryString = $this->updateCreditCardViaTr(
-            [],
-            [
-                'paymentMethodToken' => $card->token,
-                'creditCard' => [
-                    'billingAddress' => [
-                        'streetAddress' => '123 New St',
-                        'locality' => 'St. Louis',
-                        'region' => 'MO',
-                        'postalCode' => '63119',
-                        'options' => [
-                            'updateExisting' => True
-                        ]
-                    ]
-                ]
-            ]
-        );
-        $result = Braintree\CreditCard::updateFromTransparentRedirect($queryString);
-        $this->assertTrue($result->success);
-        $card = $result->creditCard;
-        $this->assertEquals(1, sizeof(Braintree\Customer::find($customer->id)->addresses));
-        $this->assertEquals('123 New St', $card->billingAddress->streetAddress);
-        $this->assertEquals('St. Louis', $card->billingAddress->locality);
-        $this->assertEquals('MO', $card->billingAddress->region);
-        $this->assertEquals('63119', $card->billingAddress->postalCode);
     }
 
     public function testSale_createsASaleUsingGivenToken()
@@ -723,7 +576,7 @@ class CreditCardTest extends Setup
             ]
         ]);
         $creditCard = $customer->creditCards[0];
-        $this->setExpectedException('Braintree\Exception\ValidationsFailed');
+        $this->expectException('Braintree\Exception\ValidationsFailed');
         Braintree\CreditCard::saleNoValidate($creditCard->token, [
             'amount' => 'invalid'
         ]);
@@ -775,7 +628,7 @@ class CreditCardTest extends Setup
             ]
         ]);
         $creditCard = $customer->creditCards[0];
-        $this->setExpectedException('Braintree\Exception\ValidationsFailed');
+        $this->expectException('Braintree\Exception\ValidationsFailed');
         Braintree\CreditCard::creditNoValidate($creditCard->token, [
             'amount' => 'invalid'
         ]);
@@ -884,19 +737,19 @@ class CreditCardTest extends Setup
 
     public function testFind_throwsIfCannotBeFound()
     {
-        $this->setExpectedException('Braintree\Exception\NotFound');
+        $this->expectException('Braintree\Exception\NotFound');
         Braintree\CreditCard::find('invalid-token');
     }
 
     public function testFind_throwsUsefulErrorMessagesWhenEmpty()
     {
-        $this->setExpectedException('InvalidArgumentException', 'expected credit card id to be set');
+        $this->expectException('InvalidArgumentException', 'expected credit card id to be set');
         Braintree\CreditCard::find('');
     }
 
     public function testFind_throwsUsefulErrorMessagesWhenInvalid()
     {
-        $this->setExpectedException('InvalidArgumentException', '@ is an invalid credit card token');
+        $this->expectException('InvalidArgumentException', '@ is an invalid credit card token');
         Braintree\CreditCard::find('@');
     }
 
@@ -931,7 +784,7 @@ class CreditCardTest extends Setup
             "share" => true
         ]);
 
-        $this->setExpectedException('Braintree\Exception\NotFound', "not found");
+        $this->expectException('Braintree\Exception\NotFound', "not found");
         Braintree\CreditCard::fromNonce($nonce);
     }
 
@@ -949,7 +802,7 @@ class CreditCardTest extends Setup
         ]);
 
         Braintree\CreditCard::fromNonce($nonce);
-        $this->setExpectedException('Braintree\Exception\NotFound', "consumed");
+        $this->expectException('Braintree\Exception\NotFound', "consumed");
         Braintree\CreditCard::fromNonce($nonce);
     }
 
@@ -1130,7 +983,7 @@ class CreditCardTest extends Setup
             'number' => '5105105105105100',
             'expirationDate' => '05/12'
         ]);
-        $this->setExpectedException('Braintree\Exception\ValidationsFailed');
+        $this->expectException('Braintree\Exception\ValidationsFailed');
         Braintree\CreditCard::updateNoValidate($creditCard->token, [
             'number' => 'invalid',
         ]);
@@ -1173,7 +1026,7 @@ class CreditCardTest extends Setup
         ]);
         Braintree\CreditCard::find($creditCard->token);
         Braintree\CreditCard::delete($creditCard->token);
-        $this->setExpectedException('Braintree\Exception\NotFound');
+        $this->expectException('Braintree\Exception\NotFound');
         Braintree\CreditCard::find($creditCard->token);
     }
 
@@ -1229,30 +1082,6 @@ class CreditCardTest extends Setup
 
         $this->assertFalse($result->success);
         $this->assertNull($result->creditCardVerification->gatewayRejectionReason);
-    }
-
-    public function createCreditCardViaTr($regularParams, $trParams)
-    {
-        $trData = Braintree\TransparentRedirect::createCreditCardData(
-            array_merge($trParams, ["redirectUrl" => "http://www.example.com"])
-        );
-        return Test\Helper::submitTrRequest(
-            Braintree\CreditCard::createCreditCardUrl(),
-            $regularParams,
-            $trData
-        );
-    }
-
-    public function updateCreditCardViaTr($regularParams, $trParams)
-    {
-        $trData = Braintree\TransparentRedirect::updateCreditCardData(
-            array_merge($trParams, ["redirectUrl" => "http://www.example.com"])
-        );
-        return Test\Helper::submitTrRequest(
-            Braintree\CreditCard::updateCreditCardUrl(),
-            $regularParams,
-            $trData
-        );
     }
 
     public function testPrepaidCard()
