@@ -10,13 +10,13 @@ class ClientTokenTest extends Setup
 {
     public function testErrorsWhenCreditCardOptionsGivenWithoutCustomerId()
     {
-        $this->setExpectedException('InvalidArgumentException', 'invalid keys: options[makeDefault]');
+        $this->expectException('InvalidArgumentException', 'invalid keys: options[makeDefault]');
         Braintree\ClientToken::generate(["options" => ["makeDefault" => true]]);
     }
 
     public function testErrorsWhenInvalidArgumentIsSupplied()
     {
-        $this->setExpectedException('InvalidArgumentException', 'invalid keys: customrId');
+        $this->expectException('InvalidArgumentException', 'invalid keys: customrId');
         Braintree\ClientToken::generate(["customrId" => "1234"]);
     }
 }
